updated model to have double as type for insurance rate

+++ CoolBlue.Products.Application/Insurance/Models/InsuranceViewModel.cs
@@ -2,6 +2,6 @@
 {
     public class InsuranceViewModel
     {
-        public float InsuranceValue { get; set; }
+        public double InsuranceValue { get; set; }
     }
 }
